% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_alignments.R
\name{add_seq_to_weighted_sequence}
\alias{add_seq_to_weighted_sequence}
\title{add_seq_to_weighted_sequence}
\usage{
add_seq_to_weighted_sequence(alignedSeq, alignedWSeq, n)
}
\arguments{
\item{alignedSeq}{Aligned Sequence}

\item{alignedWSeq}{Aligned Weighted Sequence}

\item{n}{The number of sequences in the weighted sequence}
}
\value{
Adds an aligned Sequence to a weighted sequence
}
\description{
Adds a sequence to a weighted sequence
}
\examples{
add_seq_to_weighted_sequence(alignedSeq,alignedWSeq,n)
}
