% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_aggregate.R
\name{aggregate_sequences}
\alias{aggregate_sequences}
\title{Aggregation functions}
\usage{
aggregate_sequences(unaggregated_data, format = "\%m-\%d-\%Y",
  calendar = FALSE, unit = "week", n_units = 1, anchor_table = NA,
  anchor_vector = NA, base_date = NA, occurence = min, multiset = FALSE,
  include_date = FALSE, summary_stats = TRUE)
}
\arguments{
\item{unaggregated_data}{A dataframe that has exactly 3 columns
in this order - id, date, event}

\item{format}{String specifying format of the date}

\item{calendar}{Boolean indicating whether or not to use calendar aggregation.
Defaults to false}

\item{unit}{String specifying unit of aggregation. Takes one of c("day", "week",
"month", "6 months", "year")}

\item{n_units}{Integer specifying number of units.}

\item{anchor_table}{Beta}

\item{anchor_vector}{Beta}

\item{base_date}{Beta}

\item{occurence}{Beta}

\item{multiset}{Beta}

\item{include_date}{Beta}

\item{summary_stats}{Logical controlling printing of summary
statistics regarding aggregation. Defaults to TRUE}
}
\value{
Aggregated dataframe that has sequence id, itemset (period) and event (item)
}
\description{
A dataframe having id, date and event column
can be aggregated using this function
}
\examples{
mvad \%>\%
aggregate_sequences(format = "\%Y-\%m-\%d", unit = "month", n_units = 6)
}
