% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_alignments.R
\name{align_multiple_sequences}
\alias{align_multiple_sequences}
\title{align_multiple_sequences}
\usage{
align_multiple_sequences(seqList)
}
\arguments{
\item{seqList}{A list of sequences}
}
\value{
Returns the weighted sequence after aligning all sequences in the list
}
\description{
Aligns multiple sequences making use of the align functions
}
\examples{
align_multiple_sequences(seqList)
}
