% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_alignments.R
\name{align_seq_to_Wseq}
\alias{align_seq_to_Wseq}
\title{align_seq_to_Wseq}
\usage{
align_seq_to_Wseq(seq, weightedSeq, fun = calculate_seq_soren_distance())
}
\arguments{
\item{seq}{Sequence}

\item{weightedSeq}{Weighted Sequence}

\item{fun}{The function to calculate the replacements costs based on}
}
\value{
Returns the new weighted sequence after aligning a sequence to a weighted sequence
}
\description{
Returns the new weighted sequence after aligning a sequence to a weighted sequence
}
\examples{
align_two_sequences(seq1, seq2, calculate_sorenson_distance)
}
