% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_alignments.R
\name{align_two_sequences}
\alias{align_two_sequences}
\title{align_two_sequences}
\usage{
align_two_sequences(seq1, seq2, fun = calculate_sorenson_distance)
}
\arguments{
\item{seq1}{Aligned Sequence}

\item{seq2}{Aligned Weighted Sequence}

\item{fun}{The function to calcuate the replacement sequence}
}
\value{
Returns the weighted sequence of 2 sequences after performing pairwise alignment
}
\description{
Performs pairwise alignment between 2 sequences
}
\examples{
align_two_sequences(seq1, seq2, calculate_sorenson_distance)
}
