% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_ApproxMap.R
\name{approxmap}
\alias{approxmap}
\title{approxmap}
\usage{
approxmap(file, k, cons_cutoff = 0.5, var_thresh = 0.2, noise_thresh = 0,
  period1 = "1 Week", st.date = "Mon", results_directory = "~")
}
\description{
approxmap
}
