% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-init.R
\docType{package}
\name{approxmapR}
\alias{approxmapR}
\alias{approxmapR-package}
\title{approxmapR}
\description{
Implementation of approxmap algorithm - a sequential data mining package
}
\author{
Gurudev Ilangovan <ilangurudev@gmail.com>
}
