% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_clustering.R
\name{calculate_DensityInfo}
\alias{calculate_DensityInfo}
\title{calculate_DensityInfo}
\usage{
calculate_DensityInfo(inter_seq_dist_mat, k)
}
\arguments{
\item{k}{k nearest neighbours}

\item{inter_Seq_Distance_Matrix}{A matrix of inter sequence distances}
}
\value{
Returns the density information for all the sequences
}
\description{
Calculates the density info of all the sequences
}
\examples{
calculate_DensityInfo(sequence_list, 2)
}
