% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_calculate_dist_and_cost.R
\name{calculate_dist_btw_seq_Wseq}
\alias{calculate_dist_btw_seq_Wseq}
\title{calculate_dist_btw_seq_Wseq}
\usage{
calculate_dist_btw_seq_Wseq(seq, Wseq, fun = calculate_seq_soren_distance)
}
\arguments{
\item{seq}{A sequence}

\item{Wseq}{A weighted sequence}

\item{fun}{function to calculate the costs}
}
\value{
Retuns a list of the distance matrix and the distance
}
\description{
Calculates the distance matrix between a sequence and a weighted sequence
}
\examples{
calculate_dist_btw_seq_Wseq(seq[[1]],w_seq[[2]],calculate_seq_soren_distance)
}
