% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_calculate_dist_and_cost.R
\name{calculate_dist_btw_sequences}
\alias{calculate_dist_btw_sequences}
\title{calculate_dist_btw_sequences}
\usage{
calculate_dist_btw_sequences(seq1, seq2, fun = calculate_sorenson_distance)
}
\arguments{
\item{seq1}{1st sequence}

\item{seq2}{2nd sequence}

\item{fun}{function to calculate the costs}
}
\value{
Retuns a list of the distance matrix and the distance
}
\description{
Calculates the distance matrix between 2 sequences
}
\examples{
calculate_dist_btw_sequences(seq[[1]],seq[[2]],calculate_sorenson_distance)
}
