% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_distance_bw_sequences_cpp}
\alias{calculate_distance_bw_sequences_cpp}
\title{calculate_distance_bw_sequences_cpp}
\usage{
calculate_distance_bw_sequences_cpp(sequences)
}
\arguments{
\item{sequences}{a list of intertime values}
}
\description{
Calculates x
}
\details{
\code{session_count} takes a vector of intertime values (generated via \code{\link{intertimes}},
or in any other way you see fit) and returns the total number of sessions within that dataset.
It's implimented in C++, providing a (small) increase in speed over the R equivalent.
}
