% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_calculate_dist_and_cost.R
\name{calculate_indel}
\alias{calculate_indel}
\title{calculate_indel}
\usage{
calculate_indel(itemset, fun)
}
\arguments{
\item{itemset}{Itemset from a sequence}

\item{fun}{function to calculate the indel cost}
}
\value{
Retuns the indel cost for an itemset based on the function supplied
}
\description{
Calculates the indel cost for an itemset based on the function supplied
}
\examples{
calculate_indel(seq[[1]],calculate_seq_soren_distance)
}
