% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_calculate_dist_and_cost.R
\name{calculate_indel_wseq}
\alias{calculate_indel_wseq}
\title{calculate_indel_wseq}
\usage{
calculate_indel_wseq(Witemset, fun)
}
\arguments{
\item{fun}{function to calculate the indel cost}

\item{itemset}{Itemset from a weighted sequence}
}
\value{
Retuns the indel cost for an itemset in a weighted sequence based on the function supplied
}
\description{
Calculates the indel cost for an itemset in a weighted sequence based on the function supplied
}
\examples{
calculate_indel(W-seq[[1]],calculate_seq_soren_distance)
}
