% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_calculate_dist_and_cost.R
\name{calculate_repl_btw_itemset_Witemset}
\alias{calculate_repl_btw_itemset_Witemset}
\title{calculate_repl_btw_itemset_Witemset}
\usage{
calculate_repl_btw_itemset_Witemset(itemset, Witemset, fun, n)
}
\arguments{
\item{fun}{function to calculate the replacement cost}

\item{n}{the number of strings aligned in the weighted sequence}

\item{seqItemset}{Itemset from a sequence}

\item{WseqItemset}{Itemset from a weighted sequence}
}
\value{
Retuns the replacement cost between an itemset in a sequence and an itemset in a weighted sequence using using the function supplied
}
\description{
Calculates the replacement cost between an itemset in a sequence and an itemset in a weighted sequence using using the function supplied
}
\examples{
calculate_repl_btw_itemset_Witemset(seq[[1]],Weighted_seq[[2]],calculate_seq_soren_distance,n)
}
