% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_calculate_dist_and_cost.R
\name{calculate_repl_btw_itemsets}
\alias{calculate_repl_btw_itemsets}
\title{calculate_repl_btw_itemsets}
\usage{
calculate_repl_btw_itemsets(itemset1, itemset2, fun)
}
\arguments{
\item{itemset1}{Itemset from a sequence}

\item{itemset2}{Itemset from a sequence}

\item{fun}{function to calculate the cost based on}
}
\value{
Retuns the replacement cost between an itemset in 2 sequences using the function supplied
}
\description{
Calculates the replacement cost between an itemset in 2 sequences using the function supplied
}
\examples{
calculate_repl_btw_itemsets(itemset1, itemset2, calculate_sorenson_distance)
}
