% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_calculate_dist_and_cost.R
\name{calculate_seq_soren_distance}
\alias{calculate_seq_soren_distance}
\title{calculate_seq_soren_distance}
\usage{
calculate_seq_soren_distance(seqItemset, WseqItemset, n)
}
\arguments{
\item{seqItemset}{Itemset from a sequence}

\item{WseqItemset}{Itemset from a weighted sequence}

\item{n}{The number of sequences in the weighted sequence}
}
\value{
Retuns the sorenson distance between an itemsets in a sequence and an itemset in a weighted sequence
}
\description{
Calculates the sorenson distance between an itemsets in a sequence and an itemset in a weighted sequence
}
\examples{
calculate_seq_soren_distance(seq[[1]],Weighted_seq[[2]])
}
