% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_calculate_dist_and_cost.R
\name{calculate_sorenson_distance}
\alias{calculate_sorenson_distance}
\title{calculate_sorenson_distance}
\usage{
calculate_sorenson_distance(itemset1, itemset2)
}
\arguments{
\item{itemset1}{An itemset from a sequence}

\item{itemset2}{An itemset from a sequence}
}
\value{
Returns the sorenson distance between 2 itemsets
}
\description{
Calculates the sorenson distance between 2 itemsets based on the formula
}
\examples{
calculate_sorenson_distance(sequence1[[1]],sequence1[[2]])
}
