% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_sorenson_distance_cpp}
\alias{calculate_sorenson_distance_cpp}
\title{calculate_sorenson_distance_cpp}
\usage{
calculate_sorenson_distance_cpp(itemset_1, itemset_2)
}
\arguments{
\item{itemset_1}{a vector of intertime values}

\item{itemset_2}{a vector of intertime values}
}
\description{
Calculates x
}
\details{
\code{session_count} takes a vector of intertime values (generated via \code{\link{intertimes}},
or in any other way you see fit) and returns the total number of sessions within that dataset.
It's implimented in C++, providing a (small) increase in speed over the R equivalent.
}
