% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_clustering.R
\name{cluster_knn}
\alias{cluster_knn}
\title{knnCluster}
\usage{
cluster_knn(sequences, k, id)
}
\arguments{
\item{sequences}{A list of sequences}

\item{k}{number of nearest neighbours to look at}

\item{id}{Id of every sequence in the list}

\item{seqList}{A list of sequences}

\item{k}{number of nearest neighbours to look at}

\item{id}{Id of every sequence in the list}
}
\value{
Returns the id, cluster, cluster_density and the sequence_density of each sequence

Returns the id, cluster, cluster_density and the sequence_density of each sequence
}
\description{
Performs the knn clustering algorithm and returns clustering info
}
\examples{
knnCluster(seqList,k,id)
cluster_knn

Performs the knn clustering algorithm and returns clustering info

cluster_knn(seqList,k,id)
}
