% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Preprocessing.R
\name{cvt_seq}
\alias{cvt_seq}
\title{cvt_seq}
\usage{
cvt_seq(data2)
}
\arguments{
\item{data1}{Data in the form c(id,date,item)}

\item{pd}{The period for which the data needs aggregation}

\item{st.date}{The start date from which we aggregate}
}
\value{
Returns the aggregated data in a list form usable by the rest of the program
}
\description{
Converts the data into an input usable by the other parts of the program. Calls the ord_date function.
}
\examples{
ord_date(data1,3)
}
