% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_bw_sequences_cpp}
\alias{dist_bw_sequences_cpp}
\title{dist_bw_sequences_cpp}
\usage{
dist_bw_sequences_cpp(sequence_1, sequence_2)
}
\arguments{
\item{sequence_1}{a vector of intertime values}

\item{sequence_2}{a vector of intertime values}
}
\description{
Calculates x
}
\details{
\code{session_count} takes a vector of intertime values (generated via \code{\link{intertimes}},
or in any other way you see fit) and returns the total number of sessions within that dataset.
It's implimented in C++, providing a (small) increase in speed over the R equivalent.
}
