% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_aggregate.R
\name{format_sequence}
\alias{format_sequence}
\title{Print methods}
\usage{
format_sequence(x, ...)
}
\arguments{
\item{x}{A sequence,w_sequence, w_sequence_list,
w_sequence_pattern or a w_sequence_dataframe object}

\item{...}{Additional parameters to the function that is invoked}
}
\description{
A generic function that print methods use to display formatted results.
}
