% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_aggregate.R
\name{format_sequence.Sequence}
\alias{format_sequence.Sequence}
\title{Print methods}
\usage{
\method{format_sequence}{Sequence}(sequence)
}
\arguments{
\item{sequence}{A Sequence object}
}
\value{
Returns the sequence as a formatted string
}
\description{
Print methods
}
\examples{
sequences <- pre_agg_demo \%>\% pre_aggregated() \%>\%
    convert_to_sequence() \%>\% pull(sequence)
format_sequence(sequences[[1]])
}
