% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_ApproxMap.R
\name{get_approxMap}
\alias{get_approxMap}
\title{get_approxMap}
\usage{
get_approxMap(seqList, k, strength, id = 1)
}
\arguments{
\item{seqList}{A list of sequences}

\item{k}{The number of nearest neighbours to look at}

\item{strength}{The cutoff to extract the consensus pattern from the weighted sequence}
}
\value{
Clusters and gives the consensus pattern from the sequence list
}
\description{
Does al the steps in approxmap algorithm
}
\examples{
get_approxMap(sequences_list, 2, 0.4)
}
