% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1_formatting_output.R
\name{get_C_Itemset_Formatted}
\alias{get_C_Itemset_Formatted}
\title{get_C_Itemset_Formatted}
\usage{
get_C_Itemset_Formatted(C_itemset, no_white_space = T)
}
\arguments{
\item{C_itemset}{An itemset from the consensus pattern}

\item{no_white_space}{A boolean for specifying whether or not the output should be spaced}
}
\value{
Returns the formatted itemset as a string
}
\description{
Formats an consensus itemset for displaying on the console
}
\examples{
get_C_Itemset_Formatted(C_itemset, no_white_space = T)
}
