% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1_formatting_output.R
\name{get_consensus_formatted}
\alias{get_consensus_formatted}
\title{get_consensus_formatted}
\usage{
get_consensus_formatted(consensus_pattern, no_space_bw_Items = T,
  no_space_bw_Itemsets = T)
}
\arguments{
\item{consensus_pattern}{A consensus pattern as a list}

\item{no_space_bw_Items}{A boolean for specifying whether or not the output should have space between the items}

\item{no_space_bw_Itemsets}{A boolean for specifying whether or not the output should have space between the itemsets}
}
\value{
Returns the formatted consensus pattern as a string
}
\description{
Formats a consensus pattern for displaying on the console
}
\examples{
get_consensus_formatted(consensus_pattern, no_space_bw_Items = T, no_space_bw_Itemsets = T)
}
