% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.0_shiny_helpers.R
\name{get_consensus_formatted_HTML_tablerow}
\alias{get_consensus_formatted_HTML_tablerow}
\title{get_consensus_formatted_HTML_tablerow}
\usage{
get_consensus_formatted_HTML_tablerow(weighted_seq, strength,
  blank_if_absent = F, pattern = "Consensus")
}
\description{
get_consensus_formatted_HTML_tablerow
}
