% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_ApproxMap.R
\name{get_consensus_pattern}
\alias{get_consensus_pattern}
\title{get_consensus_pattern}
\usage{
get_consensus_pattern(weighted_seq, strength, blank_if_absent = F)
}
\arguments{
\item{weighted_seq}{A weighted sequence}

\item{strength}{The cutoff to extract the consensus pattern from the weighted sequence}
}
\value{
Returns the consensus pattern if you give the weighted sequence
}
\description{
Gets the consensus pattern from the weighted sequence
}
\examples{
get_consensus_pattern(weighted_seq, 0.4)
}
