% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_clustering.R
\name{get_DensityArray}
\alias{get_DensityArray}
\title{get_DensityArray}
\usage{
get_DensityArray(DensityInfo)
}
\arguments{
\item{DensityInfo}{A list returned by calculate_DensityInfo}
}
\value{
Returns the density information for all the sequences
}
\description{
Extracts the densities of all the sequences from the density info
}
\examples{
get_DensityArray(Density_info)
}
