% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1_formatting_output.R
\name{get_Itemset_Formatted}
\alias{get_Itemset_Formatted}
\title{get_Itemset_Formatted}
\usage{
get_Itemset_Formatted(W_itemset, add_itemset_weight = T)
}
\arguments{
\item{W_itemset}{A itemset from a weighted sequence}

\item{add_itemset_weight}{A boolean specifying whether or not to add the itemset weight}
}
\value{
Returns the formatted itemset as a string
}
\description{
Formats an weighted itemset for displaying on the console
}
\examples{
get_Itemset_Formatted(W_itemset, add_itemset_weight = T)
}
