% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_aggregate.R
\name{get_n_days}
\alias{get_n_days}
\title{aggregation functions}
\usage{
get_n_days(unit, n_units)
}
\arguments{
\item{unit}{The unit of aggregation}

\item{n_units}{The number of units to aggregate}
}
\value{
Number of days to aggregate
}
\description{
A function used by \code{\link{aggregate_sequences}} to get
the unit of aggregation
}
\examples{
\\code{\\link{get_n_days("week", 2)}}
}
