% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_alignments.R
\name{get_weighted_sequence}
\alias{get_weighted_sequence}
\title{get_weighted_sequence}
\usage{
get_weighted_sequence(seq1, seq2)
}
\arguments{
\item{seq1}{Aligned Sequence 1}

\item{seq2}{Aligned Sequence 2}
}
\value{
Returns the weighted sequences for two alignments
}
\description{
Gets the weighted sequence when 2 alignments are passed
}
\examples{
get_weighted_sequence(seq1_after_aligning,seq2_after_aligning)
}
