% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1_formatting_output.R
\name{get_Wseq_Formatted}
\alias{get_Wseq_Formatted}
\title{get_Wseq_Formatted}
\usage{
get_Wseq_Formatted(W_seq, add_itemset_weight = T, no_white_space = T)
}
\arguments{
\item{W_seq}{A consensus pattern as a list}

\item{add_itemset_weight}{A boolean for specifying whether or not the output should have itemset weights}

\item{no_white_space}{A boolean for specifying whether or not the output should be spaced}
}
\value{
Returns the formatted sequence as a string
}
\description{
Formats a weighted sequence for displaying on the console
}
\examples{
get_Wseq_Formatted(W_seq, add_itemset_weight = T, no_white_space = T)
}
