% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.3_shiny_gui.R
\name{gui_approxmap}
\alias{gui_approxmap}
\title{Shiny interface for approxmap}
\usage{
gui_approxmap(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{Option will be passed on to \code{\link[shiny]{runApp}}}
}
\description{
This function calls a shiny interface for approxmap.
}
