% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_calculate_dist_and_cost.R
\name{insert_blankItemset_before}
\alias{insert_blankItemset_before}
\title{insert_blankItemset_before}
\usage{
insert_blankItemset_before(sequence)
}
\arguments{
\item{sequence}{A sequence}
}
\value{
Returns a sequence with a blank space inserted before
}
\description{
Inserts a blank space before the sequence given
}
\examples{
insert_blankItemset_before(sequence)
}
