% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_calculate_dist_and_cost.R
\name{insert_blankItemset_before_Wseq}
\alias{insert_blankItemset_before_Wseq}
\title{insert_blankItemset_before_Wseq}
\usage{
insert_blankItemset_before_Wseq(Wseq)
}
\arguments{
\item{Wseq}{A weighted Sequence}
}
\value{
Returns a weighted seq with a preceding underscore
}
\description{
Inserts a blank before the given weighted sequence
}
\examples{
insert_blankItemset_before_Wseq(weighted_seq)
}
