% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_calculate_dist_and_cost.R
\name{insert_itemset_before}
\alias{insert_itemset_before}
\title{insert_itemset_before}
\usage{
insert_itemset_before(itemset, sequence)
}
\arguments{
\item{itemset}{An itemset from a sequence}

\item{sequence}{A sequence}
}
\value{
Returns the sequence after inserting the itemset before the given sequence
}
\description{
Inserts the itemset before the given sequence
}
\examples{
insert_itemset_before(itemset,sequence)
}
