% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_calculate_dist_and_cost.R
\name{insert_itemset_before_Wseq}
\alias{insert_itemset_before_Wseq}
\title{insert_itemset_before_Wseq}
\usage{
insert_itemset_before_Wseq(weightedItemset, alignedWSeq)
}
\arguments{
\item{weightedItemset}{A weighted itemset}

\item{alignedWSeq}{An aligned weighted sequence}
}
\value{
Returns the newly weighted sequence after inserting the specified weighted itemset before the given aligned weighted sequence
}
\description{
Inserts the specified weighted itemset before the given aligned weighted sequence
}
\examples{
insert_itemset_before_Wseq(weighted_seq[[1]],alignedWSeq)
}
