% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_clustering.R
\name{inter_seq_Distance}
\alias{inter_seq_Distance}
\title{get_weighted_sequence}
\usage{
inter_seq_Distance(seqList)
}
\arguments{
\item{seqList}{A list of sequences}
}
\value{
Returns the inter-sequence distance atrix of all the sequences
}
\description{
Calculates the inter-sequence distance matrix
}
\examples{
inter_seq_Distance(sequence_list)
}
