% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_Preprocessing.R
\name{ord_date}
\alias{ord_date}
\title{ord_date}
\usage{
ord_date(data1, pd, st.date = 1)
}
\arguments{
\item{data1}{Data in the form c(id,date,item)}

\item{pd}{The period for which the data needs aggregation}

\item{st.date}{The start date from which we aggregate}
}
\value{
Returns the aggregated data in the form c(id,period,item) based on the period and the start date specified
}
\description{
Aggregates the data given in the form c(id,date,item) into periods based on the period and the start date specified
}
\examples{
ord_date(data1,3)
}
