% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.0_shiny_helpers.R
\name{plot_frequency}
\alias{plot_frequency}
\title{plot_frequency}
\usage{
plot_frequency(weighted_seq, cons_threshhold = 0.5, noise_threshold = 0,
  variation_threshold = 0.2)
}
\description{
plot_frequency
}
