% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_aggregate.R
\name{pre_aggregated}
\alias{pre_aggregated}
\title{Aggregation functions}
\usage{
pre_aggregated(df, summary_stats = TRUE)
}
\arguments{
\item{df}{The preaggregated dataframe}

\item{summary_stats}{Logical controlling printing of summary
statistics regarding aggregation. Defaults to TRUE}
}
\value{
Returns a dataframe that has the properly classes dataframe
}
\description{
Used whenever the df to be analyzed in preaggregated i.e.
instead of dates, we have periods (corresponding to itemsets)
}
\examples{
pre_agg_demo \%>\% pre_aggregated()
}
