% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_aggregate.R
\name{print.Sequence_List}
\alias{print.Sequence_List}
\title{Print methods}
\usage{
\method{print}{Sequence_List}(sequences)
}
\arguments{
\item{sequence}{A Sequence object}
}
\value{
Null. Just prints the sequence as a formatted string
using the format_sequence method
}
\description{
Print methods
}
\examples{
sequences <- pre_agg_demo \%>\% pre_aggregated() \%>\%
    convert_to_sequence()
sequences <- sequences$sequence
print(sequences)
}
