% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_aggregate.R
\name{print_raw}
\alias{print_raw}
\title{Print methods}
\usage{
print_raw(obj)
}
\arguments{
\item{obj}{Any object (useful for s3 classes)}
}
\description{
Just a method to override default print and print the raw data structure.
Can be used to examine what it looks like internally
}
\examples{
sequences <- pre_agg_demo \%>\% pre_aggregated() \%>\%
    convert_to_sequence() \%>\% pull(sequence)
print_raw(sequences[[1]])
}
