% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_distance.R
\name{repl}
\alias{repl}
\title{Distance Calculation}
\usage{
repl(x, ...)
}
\arguments{
\item{x}{Either a sequence or weighted_sequence itemset}

\item{...}{Has the other sequence itemset and function to be passed to calcaute costs}
}
\value{
Returns the replacement costs between the 2 itemsets
}
\description{
Used to calculate replacement costs.
Passes the supplied function to the subsequent functions.
}
