% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_distance.R
\name{repl.W_Sequence_Itemset}
\alias{repl.W_Sequence_Itemset}
\title{Title}
\usage{
\method{repl}{W_Sequence_Itemset}(w_sequence_itemset, sequence_itemset, n,
  fun = sorenson_distance)
}
\arguments{
\item{w_sequence_itemset}{The weighted sequence itemset}

\item{sequence_itemset}{The sequence itemset}

\item{n}{The number of sequences the weighted sequence from which the
weighted sequence itemset, is a summary of.}

\item{fun}{The distance measure to be to be used}
}
\description{
Title
}
