% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_distance.R
\name{sorenson_distance}
\alias{sorenson_distance}
\title{Distance Calculation}
\usage{
sorenson_distance(x, ...)
}
\arguments{
\item{x}{A sequence or weighted sequence itemset}

\item{...}{Contains the other itemset and other parameters to be passed
to the invoked parameter}
}
\value{
Returns the distance between the 2 itemsets
}
\description{
A genric that helps calculate sorenson distance between two itemsets
(either both from a Sequence or 1 from a sequence and another from a
weighted sequence)
}
