% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_distance.R
\name{sorenson_distance.Sequence_Itemset}
\alias{sorenson_distance.Sequence_Itemset}
\title{Distance Calculation}
\usage{
\method{sorenson_distance}{Sequence_Itemset}(sequence_itemset_1,
  sequence_itemset_2)
}
\arguments{
\item{sequence_itemset_1}{The first sequence itemset}

\item{sequence_itemset_2}{The second sequence itemset}
}
\value{
Returns the distance between the 2 itemsets
}
\description{
Distance Calculation
}
