% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_distance.R
\name{sorenson_distance.W_Sequence_Itemset}
\alias{sorenson_distance.W_Sequence_Itemset}
\title{Distance Calculation}
\usage{
\method{sorenson_distance}{W_Sequence_Itemset}(w_sequence_itemset,
  sequence_itemset, n)
}
\arguments{
\item{w_sequence_itemset}{The weighted sequence itemset}

\item{sequence_itemset}{The sequence itemset}
}
\value{
Returns the distance between the 2 itemsets
}
\description{
Distance Calculation
}
